Refactor main.py Code with Flask Blueprints and OOP
-Restructured the main code in main.py to use Flask blueprints for improved modularity and organization.
-Created a routes folder containing separate Python files (admin.py, authentication.py, community.py, cover.py, home.py, posts.py, users.py) to handle specific routes.
-Established an organized folder structure with distinct blueprints for various functionalities.
-Moved SQLAlchemy models into a models folder with an __init__.py file.
-Created community.html as the initial community page, utilizing the code from profile.html as a placeholder.
-Implemented the ability for users to follow each other with added JavaScript utilizing AJAX in users.html.
-Enhanced profile.html to display followers and following with a new div for user pictures.
-Updated HTML files to access the new routes established in the blueprints.
-Made styling changes with new colors for dark and light themes.
-Added a logo and a favicon.ico to the application.
-Implemented animations on the cover page, where content slides in and out of the screen when changing tabs.

This commit represents a significant refactoring and expansion of the codebase, introducing Flask blueprints, restructuring routes, enhancing user interactions, and refining the overall user interface.
