+++ config.py
@@ -14,3 +14,6 @@ DEBUG = True
 # HOST: The network interface on which the Flask development server will listen.
 # '0.0.0.0' means the server will be accessible from any IP address, making it suitable for development environments.
 HOST = "0.0.0.0"
+
+# ROWS_PER_PAGE: Number of posts loaded on feed
+ROWS_PER_PAGE = 10
