+++ templates/base.html
@@ -9,6 +9,7 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
+    <link rel="icon" href="{{ url_for('static', filename='assets/favicon.ico') }}" type="image/x-icon">
     {% block styles %}{% endblock %}
     {% block head %}{% endblock %}
 </head>
@@ -16,9 +17,9 @@
 <body>
     {% include "header.html" %}
 
-    <div class="background-container container-fluid">
+    <main class="background-container container-fluid">
         {% block content %}{% endblock %}
-    </div>
+    </main>
 
     {% include "footer.html" %}
