+++ templates/communities.html
@@ -23,7 +23,7 @@ Communities
                         <tr>
                             <th scope="row">{{ loop.index }}</th>
                             <td>
-                            <a href="{{ url_for('show_post', post_id=1) }}">
+                            <a href="{{ url_for('community.show_community', community_name=community.name) }}">
                                 {{ community.name }}
                             </a>
                             </td>
