+++ templates/users.html
@@ -25,7 +25,7 @@ Users
                         <tr>
                             <th scope="row">{{ loop.index }}</th>
                             <td>
-                            <a href="{{ url_for('show_profile', username=user.username) }}">
+                            <a href="{{ url_for('users.show_profile', username=user.username) }}">
                                 {{ user.username }}
                             </a>
                             </td>
@@ -33,7 +33,14 @@ Users
                             <td>{{ user.last_name }}</td>    
                             <td>{{ user.role }}</td> 
                             <td>{{ user.created_at }}</td> 
-                            <td>Follow User</td>   
+                            <td>
+                                <form class="follow-form" method="POST" action="{{ url_for('users.follow_user', user_id_to_follow=user.id) }}">
+                                    <button class="btn-primary follow-user"
+                                    role="button">
+                                    {% if current_user in user.followers.all() %}Unfollow{% else %}Follow{% endif %}
+                                    </button>
+                                </form>
+                            </td>   
                         </tr>
                         {% endfor %}
                     </tbody>
@@ -42,4 +49,34 @@ Users
         </div>
     </div>
 </main>
+{% endblock %}
+
+{% block scripts %}
+<script>
+    $(document).ready(function () {
+        // Intercept the form submissions
+        $(".follow-form").submit(function (event) {
+            // Prevent the default form submission
+            event.preventDefault();
+
+            // Reference the form being submitted
+            var $form = $(this);
+
+            // Perform AJAX request
+            $.ajax({
+                type: "POST",
+                url: $form.attr("action"),
+                success: function (data) {
+                    // Update the button text based on the response
+                    var buttonText = data.message.includes("unfollow") ? "Follow" : "Unfollow";
+                    $form.find(".follow-user").text(buttonText);
+                },
+                error: function (error) {
+                    console.error("Error:", error);
+                    // Handle error if needed
+                }
+            });
+        });
+    });
+</script>
 {% endblock %}
